get e focus moved to ixa_aux_funtion
